moves dropdown into common dir
